requirements.txt
numpy
scipy.stats
streamlit
numpy
pandas